refactor(device): extract shared AvdContextMenu; remove copy-paste duplication
Made-with: Cursor
